+++ src/main/java/net/frozenblock/wilderwild/mixin/snowlogging/PinkPetalsBlockMixin.java
@@ -47,7 +47,7 @@ public abstract class PinkPetalsBlockMixin {
 
 	@Inject(method = "createBlockStateDefinition", at = @At(value = "TAIL"))
 	public void wilderWild$createBlockStateDefinition(StateDefinition.Builder<Block, BlockState> builder, CallbackInfo info) {
-		if (!BlockConfig.canSnowlog()) return;
+		if (!BlockConfig.get().snowlogging.snowlogging) return;
 		builder.add(SnowloggingUtils.SNOW_LAYERS);
 	}