+++ src/main/java/net/frozenblock/wilderwild/mixin/snowlogging/SpreadingSnowyDirtBlockMixin.java
@@ -20,6 +20,7 @@ package net.frozenblock.wilderwild.mixin.snowlogging;
 
 import com.llamalad7.mixinextras.injector.wrapoperation.Operation;
 import com.llamalad7.mixinextras.injector.wrapoperation.WrapOperation;
+import com.llamalad7.mixinextras.sugar.Local;
 import net.frozenblock.wilderwild.block.impl.SnowloggingUtils;
 import net.minecraft.core.BlockPos;
 import net.minecraft.world.level.LevelReader;
@@ -43,10 +44,12 @@ public class SpreadingSnowyDirtBlockMixin {
 			ordinal = 0,
 			shift = At.Shift.BEFORE
 		),
-		cancellable = true,
-		locals = LocalCapture.CAPTURE_FAILHARD
+		cancellable = true
 	)
-	private static void wilderWild$canBeGrassFirstCheck(BlockState state, LevelReader levelReader, BlockPos pos, CallbackInfoReturnable<Boolean> info, BlockPos blockPos, BlockState blockState) {
+	private static void wilderWild$canBeGrassFirstCheck(
+		BlockState state, LevelReader levelReader, BlockPos pos, CallbackInfoReturnable<Boolean> info,
+		@Local(ordinal = 1) BlockState blockState
+	) {
 		if (SnowloggingUtils.isSnowlogged(blockState)) {
 			info.setReturnValue(true);
 		}
@@ -54,7 +57,10 @@ public class SpreadingSnowyDirtBlockMixin {
 
 	@WrapOperation(
 		method = "randomTick",
-		at = @At(value = "INVOKE", target = "Lnet/minecraft/world/level/block/state/BlockState;is(Lnet/minecraft/world/level/block/Block;)Z"),
+		at = @At(
+			value = "INVOKE",
+			target = "Lnet/minecraft/world/level/block/state/BlockState;is(Lnet/minecraft/world/level/block/Block;)Z"
+		),
 		slice = @Slice(
 			from = @At(
 				value = "INVOKE",