+++ src/main/java/net/frozenblock/wilderwild/mixin/snowlogging/SugarCaneBlockMixin.java
@@ -84,7 +84,7 @@ public class SugarCaneBlockMixin extends Block {
 
 	@Inject(method = "createBlockStateDefinition", at = @At(value = "TAIL"))
 	public void wilderWild$createBlockStateDefinition(StateDefinition.Builder<Block, BlockState> builder, CallbackInfo info) {
-		if (!BlockConfig.canSnowlog()) return;
+		if (!BlockConfig.get().snowlogging.snowlogging) return;
 		builder.add(SnowloggingUtils.SNOW_LAYERS);
 	}