+++ src/main/java/net/frozenblock/wilderwild/mixin/snowlogging/worldgen/ServerLevelMixin.java
@@ -34,7 +34,8 @@ import org.spongepowered.asm.mixin.injection.At;
 @Mixin(ServerLevel.class)
 public class ServerLevelMixin {
 
-	@WrapOperation(method = "tickPrecipitation",
+	@WrapOperation(
+		method = "tickPrecipitation",
 		at = @At(
 			value = "INVOKE",
 			target = "Lnet/minecraft/world/level/block/state/BlockState;is(Lnet/minecraft/world/level/block/Block;)Z",
@@ -51,21 +52,23 @@ public class ServerLevelMixin {
 		return runSnowlogging.get() || original.call(instance, block);
 	}
 
-	@WrapOperation(method = "tickPrecipitation",
+	@WrapOperation(
+		method = "tickPrecipitation",
 		at = @At(
 			value = "INVOKE",
 			target = "Lnet/minecraft/world/level/block/state/BlockState;getValue(Lnet/minecraft/world/level/block/state/properties/Property;)Ljava/lang/Comparable;",
 			ordinal = 0
 		)
 	)
-	public Comparable wilderWild$tickPrecipitationB(
-		BlockState instance, Property property, Operation<Comparable> original,
+	public Comparable<?> wilderWild$tickPrecipitationB(
+		BlockState instance, Property<?> property, Operation<Comparable<?>> original,
 		@Share("wilderWild$runSnowlogging") LocalBooleanRef runSnowlogging, @Share("wilderWild$snowloggedLayers") LocalIntRef snowloggedLayers
 	) {
 		return runSnowlogging.get() ? snowloggedLayers.get() : original.call(instance, property);
 	}
 
-	@WrapOperation(method = "tickPrecipitation",
+	@WrapOperation(
+		method = "tickPrecipitation",
 		at = @At(
 			value = "INVOKE",
 			target = "Lnet/minecraft/world/level/block/state/BlockState;setValue(Lnet/minecraft/world/level/block/state/properties/Property;Ljava/lang/Comparable;)Ljava/lang/Object;",
@@ -73,7 +76,7 @@ public class ServerLevelMixin {
 		)
 	)
 	public Object wilderWild$tickPrecipitationC(
-		BlockState instance, Property property, Comparable comparable, Operation<Object> original,
+		BlockState instance, Property<?> property, Comparable<?> comparable, Operation<Object> original,
 		@Share("wilderWild$runSnowlogging") LocalBooleanRef runSnowlogging, @Share("wilderWild$snowloggedLayers") LocalIntRef snowloggedLayers
 	) {
 		return original.call(instance, runSnowlogging.get() ? SnowloggingUtils.SNOW_LAYERS : property, comparable);